requirements.txt
ae-ffmpeg==1.1.2
argcomplete==3.1.1
auto-editor==23.32.1
av==10.0.0
boltons @ file:///work/ci_py311/boltons_1677685195580/work
brotlipy==0.7.0
certifi @ file:///croot/certifi_1683875369620/work/certifi
cffi @ file:///work/ci_py311/cffi_1676822533496/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
conda @ file:///croot/conda_1689269889729/work
conda-content-trust @ file:///work/ci_py311/conda-content-trust_1676851327497/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1685032319139/work/src
conda-package-handling @ file:///croot/conda-package-handling_1685024767917/work
conda_package_streaming @ file:///croot/conda-package-streaming_1685019673878/work
cryptography @ file:///croot/cryptography_1686613057838/work
idna @ file:///work/ci_py311/idna_1676822698822/work
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
libmambapy @ file:///croot/mamba-split_1685993156657/work/libmambapy
numpy==1.25.2
packaging @ file:///croot/packaging_1678965309396/work
Pillow==10.0.0
pluggy @ file:///work/ci_py311/pluggy_1676822818071/work
prompt-toolkit==3.0.39
pycosat @ file:///work/ci_py311/pycosat_1676838522308/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///croot/pyopenssl_1678965284384/work
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
PyYAML==6.0.1
requests @ file:///croot/requests_1682607517574/work
ruamel.yaml @ file:///work/ci_py311/ruamel.yaml_1676838772170/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
tomlkit==0.12.1
toolz @ file:///work/ci_py311/toolz_1676827522705/work
tqdm @ file:///croot/tqdm_1679561862951/work
urllib3 @ file:///croot/urllib3_1686163155763/work
wcwidth==0.2.6
xmltodict==0.13.0
zstandard @ file:///work/ci_py311_2/zstandard_1679339489613/work
